Move TextPointer and TextRange test case to fit package structure
